Work: Failed Email code and examples
Demonstrates proficiency in Regex. As part of the fundraising process we send out emails to donors and prospects; some percentage of those emails do not reach their destination and the server sends back a message detailing why an email failed. These emails need to be removed from our database, but the format of the failed email message and large volume of messages makes this a difficult and lengthy task to do manually. In response, I created this script which takes .txt file as input and returns a .csv of each unique failed email address with the reason the address did not receive our email.